html/dom/documents/resource-metadata-management/document-lastModified-01.html

<!DOCTYPE html>
<title>document.lastModified should return current time</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var t = async_test("If the last modification date of the document is not known, the current time should be returned instead.");
  var previousSeconds = document.lastModified.substring(17,20);
  setTimeout(function() {
    t.step(function() {
      assert_not_equals(document.lastModified.substring(17,20), previousSeconds);
      t.done();
    });
  }, 1000);
</script>
